+++ fibre_io.h
@@ -1,8 +1,10 @@
 #ifndef _FIBRE_IO_H
 #define _FIBRE_IO_H
 
-extern int fibre_write(int fd, const char *buf, unsigned long n);
-extern int fibre_read(int fd, char *buf, unsigned long n);
+#include <sys/types.h>
+
+extern ssize_t fibre_write(int fd, const char *buf, unsigned long n);
+extern ssize_t fibre_read(int fd, char *buf, unsigned long n);
 extern int fibre_accept(int fd);
 extern int fibre_sleep(long ms);
